% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{calColAUC}
\alias{calColAUC}
\title{Calculate AUCs for stats in colume-wise.}
\usage{
calColAUC(stats, c1, c2)
}
\arguments{
\item{stats}{matrix, ngene by n_measurements}

\item{c1}{integer vector, index of genes for condition 1}

\item{c2}{integer vector, index of genes for condition -1}
}
\value{
numeric vector, within \link{0,1}, AUCs for different measurements.
}
\description{
Calculate AUCs for stats in colume-wise.
}
