% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{evalDeltaMean}
\alias{evalDeltaMean}
\title{Evaluate delta mean between two conditions.}
\usage{
evalDeltaMean(mucond, twoHotVec)
}
\arguments{
\item{mucond}{3-dim arrays, nsample by ngene by ncond}

\item{twoHotVector}{integer vector, like (1,-1) or (0, 0, -1, 1, 0),
the two conditions we want to compare, one is 1 and the other is -1.}
}
\value{
numeric vector, length of ngene, delta mean for different genes.
}
\description{
Evaluate delta mean between two conditions.
}
