% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{evalTstat}
\alias{evalTstat}
\title{Evaluate t statistics and p-value between two conditions.}
\usage{
evalTstat(
  mucond,
  twoHotVec,
  alternative = "two.sided",
  paired = TRUE,
  var.equal = FALSE
)
}
\arguments{
\item{mucond}{3-dim arrays, nsample by ngene by ncond}

\item{twoHotVec}{integer vector, like (1,-1) or (0, 0, -1, 1, 0),
the two conditions we want to compare, one is 1 and the other is -1.}

\item{alternative}{string, default "two.sided"}

\item{paired}{bool, default FALSE}

\item{var.equal}{bool, default FALSE}
}
\value{
matrix, ngene by 2 (t statistics and pvalue in order)
NA will be used for t statistics when some troubles happend in t.test,
in this case, p-value is 1.0
}
\description{
Evaluate t statistics and p-value between two conditions.
}
