% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{extractDrawsAll}
\alias{extractDrawsAll}
\title{Extract all parameters' samples from MSSC2 Model.}
\usage{
extractDrawsAll(mssc2, genenms = NULL, method = "vi")
}
\arguments{
\item{mssc2}{MSSC2 object}

\item{genenms}{NULL or vector of strings.}

\item{method}{string, "vi" or "opt", default is "vi"}
}
\value{
list, samples for different parameters.
}
\description{
Extract all parameters' samples from MSSC2 Model.
}
