% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{extractGLMDraws}
\alias{extractGLMDraws}
\title{Extract draws/samples from the generalized linear model.}
\usage{
extractGLMDraws(mssc2, param, genenms = NULL, method = "opt")
}
\arguments{
\item{mssc2}{MSSC2 object}

\item{param}{string, name of the parameter we want to extract}

\item{genenms}{vector of string, length of ngene}

\item{method}{string, "opt" or "vi", default is "vi"}
}
\value{
NA (when have error) or vector/matrix (depend on param)
\itemize{
\item samples of the parameter for all the genes
}
}
\description{
Extract draws/samples from the generalized linear model.
}
