% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{extractGLMDrawsAll}
\alias{extractGLMDrawsAll}
\title{Extract all parameters' samples from generalised linear model.}
\usage{
extractGLMDrawsAll(mssc2, genenms = NULL, method = "opt")
}
\arguments{
\item{mssc2}{MSSC2 object}

\item{genenms}{NULL or vector of strings.}

\item{method}{string, "vi" or "opt", default is "opt"}
}
\value{
list, samples for different parameters.
}
\description{
Extract all parameters' samples from generalised linear model.
}
