% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParams.R
\name{fitDispersionParams}
\alias{fitDispersionParams}
\title{Get parameters for the prior of dispersion.
We assume the dispersion follows a log-Normal distribution.
NOTE previous name: est_varofr}
\usage{
fitDispersionParams(r, min_var = 4)
}
\arguments{
\item{r}{numeric vector, dispersions for genes}

\item{min_var}{scalar, lower bound of estimated var of r.}
}
\value{
numeric vector with four elements:
mean, variance, alpha, beta (param of inv-gamma distribution for var of logr)
}
\description{
Get parameters for the prior of dispersion.
We assume the dispersion follows a log-Normal distribution.
NOTE previous name: est_varofr
}
