% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParams.R
\name{fitGeneGlobalMeanPriorParams}
\alias{fitGeneGlobalMeanPriorParams}
\title{NOTE previous name: est_varofmu}
\usage{
fitGeneGlobalMeanPriorParams(mu, min_var = 4)
}
\arguments{
\item{mu}{numeric vector, mean expression for genes}

\item{min_var}{scalar, lower bound of estimated variance of umu}
}
\value{
numeric vector with four elements:
mean, variance, alpha, beta (param of inv-gamma distribution for var of mu)
}
\description{
NOTE previous name: est_varofmu
}
