% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParams.R
\name{fitGenewiseBatchPriorParams}
\alias{fitGenewiseBatchPriorParams}
\title{NOTE previous name: est_varofind}
\usage{
fitGenewiseBatchPriorParams(muind, min_varofind = 0.25, min_tau2 = 0.25)
}
\arguments{
\item{muind}{matrix, ngene by nind}

\item{min_varofind}{scalar, lower bound of estimated variance of individuals.}

\item{min_tau2}{scalar, lower bound of estimated tau2.}
}
\value{
list of two elements
\itemize{
\item est_varofind: matrix, nind by 4
\itemize{
\item mean, variance, alpha and beta (inv-gamma dist as the prior of the variance)
}
\item est_tau2: vector of 3 elements
\itemize{
\item tau2, tau2_alpha, tau2_beta (assume muinds follow a N(0.0, tau), taus is std.)
}
}
}
\description{
NOTE previous name: est_varofind
}
