% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParams.R
\name{fitGenewiseCondPriorParams}
\alias{fitGenewiseCondPriorParams}
\title{Estimate parameters of inv-gamma dist as the prior
for genewise conditional expression.
NOTE previous name: est_varofcond}
\usage{
fitGenewiseCondPriorParams(mucond, min_varofcond = 0.25)
}
\arguments{
\item{mucond}{matrix, ngene by ncond, gene expression level under conditions.}

\item{min_varofcond}{scalar, lower bound for the estimated variance.}
}
\value{
matrix, ngene by 3 params,
including variance of the gene expression under that condition,
alpha, beta params for the prior of the variance (inv-gamma distribution).
We treat initial means of the gene expressions as zeros.
}
\description{
Estimate parameters of inv-gamma dist as the prior
for genewise conditional expression.
NOTE previous name: est_varofcond
}
