% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParams.R
\name{fitGenewiseNBModel}
\alias{fitGenewiseNBModel}
\title{A summary fit function for Genewise Negative Binomial Model
NOTE previous name: fit_mgsnb}
\usage{
fitGenewiseNBModel(
  cnt,
  s,
  cond,
  ind,
  default_mu = 0,
  default_r = 20,
  min_var = 4,
  min_varofcond = 0.25,
  min_varofind = 0.25,
  min_tau2 = 0.25
)
}
\arguments{
\item{cnt}{matrix, count matrix size of ngene by ncell}

\item{s}{numeric vector, length of ncell, normalzing/scaling factor for cells}

\item{cond}{integer vector, length of ncell, which cond the cell belongs to,
start from 1.}

\item{ind}{integer vector, length of ncell, which individuals cells belong to,
start from 1.}
}
\value{
list of four elements
\itemize{
\item mgsnb: matrix, ngene by 2 + ncond + nind
\item mu: output of fitGeneGlobalMeanPriorParams
\item logr: output of fitGeneGlobalMeanPriorParams
\item cond: output of fitGenewiseCondPriorParams
\item ind: output of fitGenewiseBatchPriorParams
}
}
\description{
A summary fit function for Genewise Negative Binomial Model
NOTE previous name: fit_mgsnb
}
