% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanUtils.R
\name{getNameOfMatrixFromCmdstanr}
\alias{getNameOfMatrixFromCmdstanr}
\title{Get names of matrix in cmdstanr.}
\usage{
getNameOfMatrixFromCmdstanr(nm = "MuInd", nr, nc, l = "[", r = "]")
}
\arguments{
\item{nm}{string, name of variable in stan script}

\item{nr}{integer, nrow of the variable.}

\item{nc}{integer, ncolof the variable.}

\item{l}{string, left bracket, default "["}

\item{r}{string, right brackt, default "]"}
}
\value{
vector of string
}
\description{
Get names of matrix in cmdstanr.
}
