% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanUtils.R
\name{getNameOfVectorFromCmdstanr}
\alias{getNameOfVectorFromCmdstanr}
\title{Get names of vector in cmdstanr.}
\usage{
getNameOfVectorFromCmdstanr(nm = "MuInd", n = 10L, l = "[", r = "]")
}
\arguments{
\item{nm}{string, name of variable in stan script}

\item{n}{integer, length of the variable.}

\item{l}{string, left bracket, default "["}

\item{r}{string, right brackt, default "]"}
}
\value{
vector of string
}
\description{
Get names of vector in cmdstanr.
}
