% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{initMSSC2GenewiseParams}
\alias{initMSSC2GenewiseParams}
\title{Initialize MSSC2 Genewise parameters, which is further used for
initialize stan model parameters.}
\usage{
initMSSC2GenewiseParams(mssc2, cnt, s, cond, ind)
}
\arguments{
\item{mssc2}{MSSC2 object}

\item{cnt}{matrix, count matrix size of ngene by ncell}

\item{s}{numeric vector, length of ncell, normalzing/scaling factor for cells}

\item{cond}{integer vector, length of ncell, which cond the cell belongs to,
start from 1.}

\item{ind}{integer vector, length of ncell, which individuals cells belong to,
start from 1.}
}
\value{
MSSC2 object
}
\description{
Call fitGenewiseNBModel to initialize the parameters.
}
