% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParams.R
\name{initNBParam}
\alias{initNBParam}
\title{Get the mu (log-level mean) and r (dispersion) in Negative Binomial with scaling.
dispersion is defined as the same as in either in R or Stan.}
\usage{
initNBParam(y, s, min_r)
}
\arguments{
\item{y}{numeric vector, should be not all zeros.}

\item{s}{numeric vector, normalizing constant, should no zeros.}

\item{min_r}{numeric scalar, dispersion lower threshold.}
}
\value{
list of two-named scalar, mu and r
}
\description{
NOTE:
From observation, mean can be estimate well, but if it's below 1,
then r tends to be estimated too small.
}
