% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParams.R
\name{initNBParamWithCondBatch}
\alias{initNBParamWithCondBatch}
\title{Fit mu, mucond, muind for Negative Binomial with scaling.
No more Stan script based.
NOTE previous name: fit_mgsnb}
\usage{
initNBParamWithCondBatch(y, s, cond, ind, default_mu = 0, default_r = 20)
}
\arguments{
\item{y}{numeric vector, should be not all zeros.}

\item{s}{numeric vector, normalizing constant, should no zeros.}

\item{cond}{vector of integer, start from 1}

\item{ind}{vector of integer, start from 1}

\item{default_mu}{double, default is 0.0.}

\item{default_r}{double, default is 20.}
}
\value{
list of four-named element, mu, r, mucond, muind
mu and r are scalars.
mucond is a vector length of number of conditions, usually 2.
muind is a vector length of number of individuals.
}
\description{
Fit mu, mucond, muind for Negative Binomial with scaling.
No more Stan script based.
NOTE previous name: fit_mgsnb
}
