% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldCodes.R
\name{new_GenewiseNBFit}
\alias{new_GenewiseNBFit}
\title{Create GenewiseNBFit S3 object
It's used for MSSC model initialization.}
\usage{
new_GenewiseNBFit()
}
\value{
GenewiseNBFit object
}
\description{
Create GenewiseNBFit S3 object
It's used for MSSC model initialization.
}
