% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_InitMssc.R
\name{new_InitMssc}
\alias{new_InitMssc}
\title{Create InitMssc object}
\usage{
new_InitMssc()
}
\value{
InitMssc
}
\description{
Initialize a InitMssc object in order to initialize the parameters
of MSSC.
}
