% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParams.R
\name{normLogMean}
\alias{normLogMean}
\title{Estimate the normalized mean of a numeric vector on log-level,
which is similar with log(mean(y/s)) (or better),}
\usage{
normLogMean(y, s)
}
\arguments{
\item{y}{numeric vector, should be not all zeros.}

\item{s}{numeric vector, normalizing constant, should no zeros.}
}
\value{
numeric scalar
}
\description{
NOTE previsous name: init_snb_log_mean
}
