% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{PSIS}
\alias{PSIS}
\title{Get Pareto Smoothed Importance Sampling weights.
Weights are normalized and NOT in scale level (what do I mean?) by default.}
\usage{
PSIS(mssc2, takelog = FALSE, donormalize = TRUE)
}
\value{
list of two elements:
\itemize{
\item psis: output of loo:psis
\item weight: output of loo::weights.importance_sampling
}
}
\description{
This is specific for variational inference.
}
