% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanUtils.R
\name{repVecColwise}
\alias{repVecColwise}
\title{Replicate a vector in column-wise for n times.}
\usage{
repVecColwise(x, n)
}
\arguments{
\item{x}{vector of numeric}

\item{n}{integer, times of repeat in column-wise}
}
\value{
matrix
}
\description{
NOTE previous name: rep_col
}
