% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanUtils.R
\name{repVecRowise}
\alias{repVecRowise}
\title{Replicate a vector in row-wise for n times.}
\usage{
repVecRowise(x, n)
}
\arguments{
\item{x}{vector of numeric}

\item{n}{integer, times of repeat in row-wise}
}
\value{
matrix
}
\description{
NOTE previous name: rep_row
}
