% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{runGLMAP}
\alias{runGLMAP}
\title{Run MAP optimization for generalized lineaer model.}
\usage{
runGLMAP(mssc2, data)
}
\arguments{
\item{mssc2}{MSSC2 object}

\item{data}{list, generated by toStanInput}
}
\value{
MSSC2 object
}
\description{
Run MAP optimization for generalized lineaer model.
}
