% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{runMAP}
\alias{runMAP}
\title{Run MAP optimization for mssc2.}
\usage{
runMAP(mssc2, data)
}
\arguments{
\item{mssc2}{MSSC2 object}

\item{data}{list, generated by toStanInput}
}
\value{
MSSC2 object
}
\description{
Run MAP optimization for mssc2.
}
