% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{setGLMInitialParams}
\alias{setGLMInitialParams}
\title{Set initial values for the generalized linear model.
TODO: muind estimated by the genes from DESEQ2.}
\usage{
setGLMInitialParams(mssc2)
}
\arguments{
\item{mssc2}{MSSC2 object}
}
\value{
MSSC2 object
}
\description{
Set initial values for the generalized linear model.
TODO: muind estimated by the genes from DESEQ2.
}
