% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{setStanHyperparams}
\alias{setStanHyperparams}
\title{Set hyper parameters for the stan model of mssc2.}
\usage{
setStanHyperparams(mssc2)
}
\arguments{
\item{mssc2}{MSSC2 object}
}
\value{
MSSC2 object
}
\description{
Set hyper parameters for the stan model of mssc2.
}
