% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{setStanInitialParams}
\alias{setStanInitialParams}
\title{Set initial values for the parameters in the stan model of mssc2.}
\usage{
setStanInitialParams(mssc2)
}
\arguments{
\item{mssc2}{MSSC2 object}
}
\value{
MSSC2 object
}
\description{
TODO: muind estimated by the genes from DESEQ2.
}
