% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanUtils.R
\name{split_matrix_col}
\alias{split_matrix_col}
\title{Split matrix columns into a matrix, which is row-major order}
\usage{
split_matrix_col(mat, secondim, secondimNms = NULL)
}
\arguments{
\item{mat}{matrix, size of nsample by (second_dim * third_dim)}

\item{secondim}{integer, secondim we want}

\item{secondimNms}{vector of string, names for secondim.}
}
\value{
three-dim array
}
\description{
Split matrix columns into a matrix, which is row-major order
}
