% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssc2.R
\name{toStanInput}
\alias{toStanInput}
\title{Convert input data to the format of Stan Input.}
\usage{
toStanInput(cnt, ind, cond, s, hp)
}
\arguments{
\item{cnt}{matrix, count matrix size of ngene by ncell}

\item{ind}{integer vector, length of ncell, which individuals cells belong to,
start from 1}

\item{cond}{integer vector, length of ncell, which conditions cells belong to,
start from 1.}

\item{s}{normalizing/scaling factor, length of ncell.}

\item{hp}{list of 4 elements.}
}
\description{
Convert input data to the format of Stan Input.
}
